feat(component): Allow `Button`s to have `href`
A pretty common use case is to use a button visual
style in combination with a link to another page.

However, HTML doesn't support nesting `<a>` inside
of `<button>`.

Adding an `onClick` hook to the `<button>` is fine,
but that only works with JavaScript enabled.

In this case, they will render as an anchor (`<a>`)
instead, which is the best scenario for SSR apps.