fix(lib): proper window check in ssr/ssg pages
While rendering Modal, there was a difference between the React tree that was pre-rendered (SSR/SSG) and the React tree that rendered during the first render in the Browser (Hydration). Due to which there was issues with hydration when used in SSR/SSG. This issue was caused by the windowExists() that was using window in a component's rendering.

Signed-off-by: chankruze <chankruze@gmail.com>