feat(Avatar): Support overriding the Avatar img element
It accomplishes this by changing the `img` prop to either be a
string or a render prop that returns a React element.
That makes it so any of the props can be overridden, or an entirely
different element could be used in its place.

The docs site doesn't support the render prop syntax by default, so I
added a small extension that allows the docs pages to provide a
`rawCode` property which is passed directly to the syntax highlighter
instead of generating the code sample from the `code` JSX.

Includes a new passing test and a Storybook example.