chore(pre-commit): simplify pre-commit logic to only run on changed files
Configure the project to use lint-staged for the pre-commit logic,
instead of having lint-staged installed but not using it.

This should speed up development, and make it easier to commit changes
to the project. There is less time waiting for the whole project to
build or for all of the tests to run, when they are not related to your
changes.

The new pre-commit workflow is:
- eslint and prettier will run one after another, instead of in
  parallel. This is to prevent race conditions with potentially
modifying the same file: https://github.com/okonet/lint-staged#task-concurrency
- Recommended vitest options for lint-staged: https://vitest.dev/guide/cli.html#vitest-related
- tsc-files was added as a devDependency, in order to only run tsc
  against the changed files, instead of all files: https://github.com/gustavopch/tsc-files

Also the `lint-staged.js` file was renamed since it was not detectable
by lint-staged so it did not work. The new name is the one suggested by lint-staged for use
with the `module.exports` syntax: https://github.com/okonet/lint-staged#configuration

Fixes #671