feat(textinput.tsx): added renderIcon and renderRightIcon to allow for an Icon onClick handler
This change allows for users to render Icons using the renderIcon function property, instead of the
icon property. This gives users the ability to add custom props to their Icon, such as, an onClick
handler. The same goes for the renderRightIcon property, but the Icon will be displayed on the right
side of the text input instead.

TextInput icon onClick #734