docs(app/components/code-preview): only display function's name in code previews, for now
We would like to be able to show function bodies, but that creates an undesirable result for FCs
that are imported inside props, like `icon={<HiInformationCircle />}`, which becomes a mess. Same
with `icon={HiInformationCircle}`.  We should revisit this in the future, but for now, changing the
default to only display the name of the function is sufficient. We can probably attach
`displayName`s to external components we use to get their proper name, but there might be an even
easier solution.