fix(modal): fixed a bug that disallowed users to set the position of a modal
The modal's position property is handled by the parent of the actual content box using flex, align
items, and justify content but was being blocked by the content box's proptery of margin: auto. This
fix deletes that margin: auto property so the parent can corrently handle the position of the modal
correctly.

Placement Attribute of the Modal Component doesn't work #760