docs(app/docs.css): fix custom CSS wasn't applied to code previews
We use a variety of custom CSS hooks in traditional CSS file (not Tailwind CSS) to override the
third-party plugin we use to render code previews. The `docs.css` file was not "cascading" as
intended in production, and all of the styles we wrote there suddenly disappeared in production
mode.