docs: add Theme section to each component's docs (#815)
docs(app/docs/components/*): add Theme section to each component's docs

For now, we're just including the snippet of Tailwind CSS classes that are applied to each key in
the component's theme. That makes it easy to figure out where you need to make tweaks to the theme.
I'm not sure how we can make it more useful yet - we will want to get feedback on this.