feat(dropdown.tsx): Dropdown Accessibility (#840)
* feat(button.tsx): created BaseButton component

This component is responsible to handle the logic of creating the button element based on the custom or default tag name

* feat(dropdown.tsx): enabled Dropdown to be accessible by keyboard

Now the Dropdown component is accessible by keyboard. Also the dismiss on item click is fixed

fix #648. fix #793

* feat(dropdown.tsx): removed div wrapper around trigger button

* feat(dropdown.tsx): moved renderTrigger function to Trigger

clone element resulted from renderTrigger and attach the ref and a11y props to it

* test(dropdown.tsx): updated and added more tests

Resolution for dependency nwsapi was added in order to fix an issue in the tababble package

* refact(floating.tsx): passed arrowRef to useBaseFloating custom hook

* fix(floating.tsx): readded focus interaction

* test(dropdown.tsx): added a delay in one of the test cases

* docs(dropdown.tsx): added stories and documentation section

* refact(dropdown.tsx): removed mergeWrapperClassName utility fn

* feat(dropdown.tsx): added theme customization support for menuitem