feat(component): new pauseOnHover prop to <Carousel>, provide quick information with chance to pause
A new property on the <Carousel> component, `pauseOnHover`, will improve accessibility, allow for a
faster flow of information, and help users find desired information quicker. From the Web
Accessibility Initiative in "What makes a carousel accessible?", they note: "Users must be able to
pause carousel movement because it can be too fast or distracting, making text hard to read."
(www.w3.org/WAI/tutorials/carousels/) Currently, there is no built in feature to accomplish this.
Developers will either have to disable the `slide` property, or make the `slideInterval` property
very long to allow users to read necessary information. With the `pauseOnHover` property, users will
have a quick way, both on mobile and desktop, to pause on the desired slide. Furthermore, this
allows developers to display more information at a faster pace to users because users have a choice
at which slide to stop and read, but can still have a general understanding of information displayed
on other slides.