feat: add `<Banner>` component (#949)
* feat(src/components/banner): add component, Banner

This component is currently a work-in-progress; you can expect that there may be significant changes
in the future, in particular, to its theme. There currently isn't a theme, but there probably will
be in the future.

* docs(app/docs/components/banner): add examples for `<Banner>` component

* docs(app/docs/layout): add sidebar link to `Banner` docs

* feat(src/components/banner): add `role="banner"` to `<Banner>`

* docs(app/docs/components/banner): add missing `import`s

* docs(app/docs/components/*): fix component intro docs not displaying

We needed to move the Table of Contents declaration - which is replaced inline by a Markdown plugin
- to immediately above the first second-level heading (the first item in the Table of Contents). Now
the intro text for each component's docs is showing again!