fix(datepicker): fix the inconsistent weekday labels
Updating the WeekStart enum to be aligned with the values Javascript Date object uses allows to
simplify the code and maintain compaitbility. Fixed the off-by-one error

BREAKING CHANGE: As the WeekStart enum changed order to be aligned with Javascript Date object, now
you have to change your weekStart attribute to be -1, so for Monday you should put 1, instead of 2.
But it wasn't working before as the values were incorrectly rendered in the first place.

fix #1044